moving the lib_reloader macro into submodule
