+++ hot-lib-reloader-macro/src/lib_reloader/generate_libreloader_struct.rs
